#601: Show total Lmm count on Profile screen

+++ domain/src/main/kotlin/com/ringoid/domain/repository/feed/IFeedRepository.kt
@@ -50,7 +50,7 @@ interface IFeedRepository {
     val feedMatches: PublishSubject<List<FeedItem>>
     val feedMessages: PublishSubject<List<FeedItem>>
     val lmmChanged: PublishSubject<Boolean>
-    val lmmLoadFinish: PublishSubject<Long>
+    val lmmLoadFinish: PublishSubject<Int>
     val newLikesCount: PublishSubject<Int>
     val newMatchesCount: PublishSubject<Int>
     val newMessagesCount: PublishSubject<Int>
